ref: extract LottieAnimation component to templates dir
