+++ frontend/src/features/rollout-announcement/components/NewFeatureContent.tsx
@@ -1,40 +1,9 @@
-import { useEffect, useRef } from 'react'
-import { Box, BoxProps, ModalBody, ModalHeader, Text } from '@chakra-ui/react'
-import lottie, { AnimationConfigWithData, SVGRendererConfig } from 'lottie-web'
+import { ModalBody, ModalHeader, Text } from '@chakra-ui/react'
+import { AnimationConfigWithData } from 'lottie-web'
+
+import { LottieAnimation } from '~templates/LottieAnimation'
 
 import { NewFeatureTag } from './NewFeatureTag'
-
-interface LottieAnimationProps extends BoxProps {
-  animationData: AnimationConfigWithData['animationData']
-  preserveAspectRatio?: SVGRendererConfig['preserveAspectRatio']
-}
-
-export const LottieAnimation = ({
-  animationData,
-  preserveAspectRatio,
-  ...boxProps
-}: LottieAnimationProps): JSX.Element => {
-  const element = useRef<HTMLDivElement>(null)
-  const lottieInstance = useRef<unknown>()
-
-  useEffect(() => {
-    if (!element.current) return
-
-    lottieInstance.current = lottie.loadAnimation({
-      animationData,
-      container: element.current,
-      rendererSettings: {
-        preserveAspectRatio: preserveAspectRatio,
-      },
-    })
-
-    return () => {
-      lottieInstance.current = lottie.destroy()
-    }
-  }, [animationData, preserveAspectRatio])
-
-  return <Box {...boxProps} ref={element} />
-}
 
 interface NewFeatureContentProps {
   title: string
@@ -50,8 +19,6 @@ export const NewFeatureContent = (props: {
   return (
     <>
       <LottieAnimation
-        // The link will always change in Chromatic so this should be ignored.
-        data-chromatic="ignore"
         bg="primary.100"
         pt="4.5rem"
         height={{ base: '30vh', md: 'initial' }}
